Split classes out into separate files.
